Remove StoreContainer trait and instance checking
